requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiofiles>=23.0.0
Pillow>=10.0.0 